♻️ refactor: Refactor Image components
