+++ src/components/FileList/index.tsx
@@ -1,6 +1,6 @@
-import { memo } from 'react';
-
-import LightBox from '@/components/FileList/Lightbox';
+import { ImageGallery } from '@lobehub/ui';
+import { memo, useMemo } from 'react';
+import { Flexbox } from 'react-layout-kit';
 
 import FileItem from './FileItem';
 import { useStyles } from './style';
@@ -14,15 +14,25 @@ interface FileListProps {
 const FileList = memo<FileListProps>(({ items, editable = true, alwaysShowClose }) => {
   const { styles } = useStyles();
 
-  return (
-    <div className={styles.container}>
-      <LightBox items={items}>
+  const content = useMemo(
+    () => (
+      <ImageGallery enable={items?.length > 1}>
         {items.map((i) => (
           <FileItem alwaysShowClose={alwaysShowClose} editable={editable} id={i} key={i} />
         ))}
-      </LightBox>
-    </div>
+      </ImageGallery>
+    ),
+    [items],
   );
+
+  if (editable)
+    return (
+      <Flexbox gap={8} horizontal>
+        {content}
+      </Flexbox>
+    );
+
+  return <div className={styles.container}>{content}</div>;
 });
 
 export default FileList;
